refactor: move to a separate MacOS folder
